fixed posixpath and moved gen_exports.py
